python/algorithmia/requirements.txt
algorithmia>=1.0.0,<2.0
tensorflow
six
numpy
pillow